update 24/03
I made some updates to verify.php. The file had PHP, JS, and HTML code, and the issue was that the PHP code ran first, followed by the HTML and JS code. As a result, the database update didn't wait for the required 7 seconds and closed too quickly. I fixed this problem by removing the JS and HTML code and using only PHP code, which resolved the issue.

Next, I needed to find a way to let the user know whether or not they could open the door. I added an event listener to the submit button that removes all elements from the page, changes the background color to green or red, and refreshes the page.

Finally, I removed the testcookies.php file, as it was no longer needed.