update 25/03
"discovered" what was wrong with updating the db. Whenever i try to change the background color it broke, so it has to do with that i dont know if its because of the settimeout are something else i will just give up on this and try something different. That something different will be a LED or a something that makes sound to indicate that the door has opened.